Calculate clipPos forward strand even on reverse strand reads to get sequence track right on both top and bottom
