+++ plugins/sequence/src/DivSequenceRenderer/components/DivSequenceRendering.tsx
@@ -6,7 +6,7 @@ import { Region } from '@jbrowse/core/util/types'
 import { createJBrowseTheme } from '@jbrowse/core/ui'
 import { observer } from 'mobx-react'
 import React from 'react'
-import { bpSpanPx } from '@jbrowse/core/util'
+import { bpSpanPx, revcom, complement } from '@jbrowse/core/util'
 
 interface MyProps {
   features: Map<string, Feature>
@@ -19,56 +19,6 @@ interface MyProps {
   showReverse: boolean
   showTranslation: boolean
 }
-
-function revcom(seqString: string) {
-  return complement(seqString).split('').reverse().join('')
-}
-
-const complement = (() => {
-  const complementRegex = /[ACGT]/gi
-
-  // from bioperl: tr/acgtrymkswhbvdnxACGTRYMKSWHBVDNX/tgcayrkmswdvbhnxTGCAYRKMSWDVBHNX/
-  // generated with:
-  // perl -MJSON -E '@l = split "","acgtrymkswhbvdnxACGTRYMKSWHBVDNX"; print to_json({ map { my $in = $_; tr/acgtrymkswhbvdnxACGTRYMKSWHBVDNX/tgcayrkmswdvbhnxTGCAYRKMSWDVBHNX/; $in => $_ } @l})'
-  const complementTable = {
-    S: 'S',
-    w: 'w',
-    T: 'A',
-    r: 'y',
-    a: 't',
-    N: 'N',
-    K: 'M',
-    x: 'x',
-    d: 'h',
-    Y: 'R',
-    V: 'B',
-    y: 'r',
-    M: 'K',
-    h: 'd',
-    k: 'm',
-    C: 'G',
-    g: 'c',
-    t: 'a',
-    A: 'T',
-    n: 'n',
-    W: 'W',
-    X: 'X',
-    m: 'k',
-    v: 'b',
-    B: 'V',
-    s: 's',
-    H: 'D',
-    c: 'g',
-    D: 'H',
-    b: 'v',
-    R: 'Y',
-    G: 'C',
-  } as { [key: string]: string }
-
-  return (seqString: string) => {
-    return seqString.replace(complementRegex, m => complementTable[m] || '')
-  }
-})()
 
 const defaultStarts = ['ATG']
 const defaultStops = ['TAA', 'TAG', 'TGA']
